+++ client/web/src/enterprise/insights/components/insights-view-grid/components/backend-insight/BackendInsight.tsx
@@ -5,7 +5,7 @@ import { useMergeRefs } from 'use-callback-ref'
 
 import { asError } from '@sourcegraph/common'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
-import { useDebounce, useDeepMemo } from '@sourcegraph/wildcard'
+import { Link, useDebounce, useDeepMemo } from '@sourcegraph/wildcard'
 
 import { SeriesDisplayOptionsInput } from '../../../../../../graphql-operations'
 import { BackendInsight, CodeInsightsBackendContext, InsightFilters } from '../../../../core'
@@ -141,6 +141,8 @@ export const BackendInsightView: React.FunctionComponent<React.PropsWithChildren
         insightType: getTrackingTypeByInsightType(insight.type),
     })
 
+    const shareableUrl = `${window.location.origin}/insights/insight/${insight.id}`
+
     return (
         <InsightCard
             {...otherProps}
@@ -150,7 +152,13 @@ export const BackendInsightView: React.FunctionComponent<React.PropsWithChildren
             onMouseEnter={trackMouseEnter}
             onMouseLeave={trackMouseLeave}
         >
-            <InsightCardHeader title={insight.title}>
+            <InsightCardHeader
+                title={
+                    <Link to={shareableUrl} target="_blank" rel="noopener noreferrer">
+                        {insight.title}
+                    </Link>
+                }
+            >
                 {isVisible && (
                     <>
                         <DrillDownFiltersPopover