+++ enterprise/cmd/repo-updater/internal/authz/perms_syncer_test.go
@@ -37,7 +37,7 @@ func TestPermsSyncer_ScheduleUsers(t *testing.T) {
 	authz.SetProviders(true, []authz.Provider{&mockProvider{}})
 	defer authz.SetProviders(true, nil)
 
-	s := NewPermsSyncer(nil, nil, nil, nil, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), nil, nil, nil, nil, nil)
 	s.ScheduleUsers(context.Background(), authz.FetchPermsOptions{}, 1)
 
 	expHeap := []*syncRequest{
@@ -56,7 +56,7 @@ func TestPermsSyncer_ScheduleRepos(t *testing.T) {
 	authz.SetProviders(true, []authz.Provider{&mockProvider{}})
 	defer authz.SetProviders(true, nil)
 
-	s := NewPermsSyncer(nil, nil, nil, nil, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), nil, nil, nil, nil, nil)
 	s.ScheduleRepos(context.Background(), 1)
 
 	expHeap := []*syncRequest{
@@ -181,7 +181,7 @@ func TestPermsSyncer_syncUserPerms(t *testing.T) {
 		eauthz.MockProviderFromExternalService = nil
 	}()
 
-	s := NewPermsSyncer(db, reposStore, perms, timeutil.Now, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), db, reposStore, perms, timeutil.Now, nil)
 
 	p.fetchUserPerms = func(context.Context, *extsvc.Account) (*authz.ExternalUserPermissions, error) {
 		return &authz.ExternalUserPermissions{
@@ -266,7 +266,7 @@ func TestPermsSyncer_syncUserPerms_noPerms(t *testing.T) {
 	})
 	perms.UserIsMemberOfOrgHasCodeHostConnectionFunc.SetDefaultReturn(true, nil)
 
-	s := NewPermsSyncer(db, reposStore, perms, timeutil.Now, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), db, reposStore, perms, timeutil.Now, nil)
 
 	tests := []struct {
 		name     string
@@ -350,7 +350,7 @@ func TestPermsSyncer_syncUserPerms_tokenExpire(t *testing.T) {
 	perms := edb.NewMockPermsStore()
 	perms.UserIsMemberOfOrgHasCodeHostConnectionFunc.SetDefaultReturn(true, nil)
 
-	s := NewPermsSyncer(db, reposStore, perms, timeutil.Now, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), db, reposStore, perms, timeutil.Now, nil)
 
 	t.Run("invalid token", func(t *testing.T) {
 		p.fetchUserPerms = func(ctx context.Context, account *extsvc.Account) (*authz.ExternalUserPermissions, error) {
@@ -449,7 +449,7 @@ func TestPermsSyncer_syncUserPerms_prefixSpecs(t *testing.T) {
 	perms := edb.NewMockPermsStore()
 	perms.UserIsMemberOfOrgHasCodeHostConnectionFunc.SetDefaultReturn(true, nil)
 
-	s := NewPermsSyncer(db, reposStore, perms, timeutil.Now, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), db, reposStore, perms, timeutil.Now, nil)
 
 	p.fetchUserPerms = func(context.Context, *extsvc.Account) (*authz.ExternalUserPermissions, error) {
 		return &authz.ExternalUserPermissions{
@@ -520,7 +520,7 @@ func TestPermsSyncer_syncUserPerms_subRepoPermissions(t *testing.T) {
 	perms := edb.NewMockPermsStore()
 	perms.UserIsMemberOfOrgHasCodeHostConnectionFunc.SetDefaultReturn(true, nil)
 
-	s := NewPermsSyncer(db, reposStore, perms, timeutil.Now, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), db, reposStore, perms, timeutil.Now, nil)
 
 	p.fetchUserPerms = func(context.Context, *extsvc.Account) (*authz.ExternalUserPermissions, error) {
 		return &authz.ExternalUserPermissions{
@@ -554,7 +554,7 @@ func TestPermsSyncer_syncRepoPerms(t *testing.T) {
 	db.ReposFunc.SetDefaultReturn(mockRepos)
 
 	newPermsSyncer := func(reposStore repos.Store, perms edb.PermsStore) *PermsSyncer {
-		return NewPermsSyncer(db, reposStore, perms, timeutil.Now, nil)
+		return NewPermsSyncer(logtest.Scoped(t), db, reposStore, perms, timeutil.Now, nil)
 	}
 
 	t.Run("TouchRepoPermissions is called when no authz provider", func(t *testing.T) {
@@ -762,7 +762,7 @@ func TestPermsSyncer_syncRepoPerms(t *testing.T) {
 func TestPermsSyncer_waitForRateLimit(t *testing.T) {
 	ctx := context.Background()
 	t.Run("no rate limit registry", func(t *testing.T) {
-		s := NewPermsSyncer(nil, nil, nil, nil, nil)
+		s := NewPermsSyncer(logtest.Scoped(t), nil, nil, nil, nil, nil)
 
 		ctx, cancel := context.WithTimeout(ctx, time.Second)
 		defer cancel()
@@ -774,7 +774,7 @@ func TestPermsSyncer_waitForRateLimit(t *testing.T) {
 
 	t.Run("enough quota available", func(t *testing.T) {
 		rateLimiterRegistry := ratelimit.NewRegistry()
-		s := NewPermsSyncer(nil, nil, nil, nil, rateLimiterRegistry)
+		s := NewPermsSyncer(logtest.Scoped(t), nil, nil, nil, nil, rateLimiterRegistry)
 
 		ctx, cancel := context.WithTimeout(ctx, time.Second)
 		defer cancel()
@@ -789,7 +789,7 @@ func TestPermsSyncer_waitForRateLimit(t *testing.T) {
 		l := rateLimiterRegistry.Get("extsvc:github:1")
 		l.SetLimit(1)
 		l.SetBurst(1)
-		s := NewPermsSyncer(nil, nil, nil, nil, rateLimiterRegistry)
+		s := NewPermsSyncer(logtest.Scoped(t), nil, nil, nil, nil, rateLimiterRegistry)
 
 		ctx, cancel := context.WithTimeout(ctx, time.Second)
 		defer cancel()
@@ -810,7 +810,7 @@ func TestPermsSyncer_syncPerms(t *testing.T) {
 	}
 
 	// Request should be removed from the queue even if error occurred.
-	s := NewPermsSyncer(nil, nil, nil, nil, nil)
+	s := NewPermsSyncer(logtest.Scoped(t), nil, nil, nil, nil, nil)
 	s.queue.Push(request)
 
 	expErr := "unexpected request type: 3"
@@ -853,7 +853,7 @@ func TestPermsSyncer_maybeRefreshGitLabOAuthTokenFromAccount(t *testing.T) {
 			})
 			db.UserExternalAccountsFunc.SetDefaultReturn(externalAccounts)
 
-			s := NewPermsSyncer(db, nil, nil, timeutil.Now, nil)
+			s := NewPermsSyncer(logtest.Scoped(t), db, nil, nil, timeutil.Now, nil)
 
 			// http mocking
 			server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {